+++ package.json
@@ -19,7 +19,6 @@
     "@auth0/auth0-spa-js": "^1.13.2",
     "apollo-link-batch-http": "^1.2.14",
     "codemirror": "^5.58.2",
-    "codemirror-graphql": "^0.12.3",
     "cronstrue": "^1.103.0",
     "d3": "^6.2.0",
     "d3-dag": "0.3.5",
